analog zu 'https://www.udemy.com/spring-framework-5-beginner-to-guru/learn/lecture/7497700' (Abschnitt 10, Video 180)
Common design pattern for testing the business logic of a Service layer:
a simple example on how to create a Mock-object to test a Service-Implementation as a pure JUnit Unit test without further dependencies. We test a Spring class, but without having to use (or start) the Spring context.

CMD + EINFG öffnet Kontextmenü "Generate" in einer Klasse in Intellij IDEA

"tests" => "java" (Rightclick) => "Run 'All tests'"